separated tabs from ui crate
