+++ Cargo.toml
@@ -14,17 +14,12 @@ repository = "https://github.com/rewin123/space_editor"
 
 [workspace]
 members = [
-    "crates/prefab", 
-    "crates/shared", 
-    "crates/undo", 
-    "crates/persistence",
-    "crates/editor_core", 
-    "crates/editor_ui", 
+    "crates/*",
     "modules/bevy_xpbd_plugin"
-]
+, "crates/editor_tabs"]
 
 [workspace.package]
-version = "0.5.0"
+version = "0.6.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 authors = ["rewin <rewin1996@gmail.com>", "Julia Naomi <jnboeira@outlook.com>"]
@@ -37,12 +32,13 @@ homepage = "https://github.com/rewin123/space_editor"
 bevy = "0.13.1"
 
 # Editor Crates
-space_prefab = { version = "0.5.0", path = "crates/prefab" }
-space_shared = { version = "0.5.0", path = "crates/shared" }
-space_undo = { version = "0.5.0", path = "crates/undo" }
-space_persistence = { version = "0.5.0", path = "crates/persistence"}
-space_editor_core = { version = "0.5.0", path = "crates/editor_core", features = ["persistence_editor"] }
-space_editor_ui = { version = "0.5.0", path = "crates/editor_ui", features = ["persistence_editor"] }
+space_prefab = { version = "0.6.0", path = "crates/prefab" }
+space_shared = { version = "0.6.0", path = "crates/shared" }
+space_undo = { version = "0.6.0", path = "crates/undo" }
+space_persistence = { version = "0.6.0", path = "crates/persistence"}
+space_editor_core = { version = "0.6.0", path = "crates/editor_core", features = ["persistence_editor"] }
+space_editor_ui = { version = "0.6.0", path = "crates/editor_ui", features = ["persistence_editor"] }
+space_editor_tabs = { version = "0.6.0", path = "crates/editor_tabs" }
 
 # Crates inner libraries
 anyhow = "1.0"
@@ -74,7 +70,7 @@ ron = "0.8"
 serde = "1"
 
 # Community Modules
-space_bevy_xpbd_plugin = { version = "0.5.0", path = "modules/bevy_xpbd_plugin"}
+space_bevy_xpbd_plugin = { version = "0.6.0", path = "modules/bevy_xpbd_plugin"}
 
 [dependencies]
 bevy.workspace = true
