+++ crates/editor_ui/Cargo.toml
@@ -22,6 +22,7 @@ space_prefab.workspace = true
 space_undo.workspace = true
 space_persistence.workspace = true
 space_shared.workspace = true
+space_editor_tabs.workspace = true
 
 serde.workspace = true
 pretty-type-name.workspace = true
