+++ crates/editor_ui/src/settings.rs
@@ -4,6 +4,7 @@ use bevy::{
 };
 use bevy_egui::*;
 use space_editor_core::hotkeys::AllHotkeys;
+use space_editor_tabs::{NewTabBehaviour, NewWindowSettings};
 use space_shared::ext::bevy_inspector_egui::bevy_inspector;
 use space_undo::ChangeChainSettings;
 
@@ -16,12 +17,6 @@ use super::{
     editor_tab::{EditorTab, EditorTabName},
     EditorUiAppExt,
 };
-
-const TAB_MODES: [NewTabBehaviour; 3] = [
-    NewTabBehaviour::Pop,
-    NewTabBehaviour::SameNode,
-    NewTabBehaviour::SplitNode,
-];
 
 const GAME_MODES: [GameMode; 2] = [GameMode::Game2D, GameMode::Game3D];
 
@@ -109,48 +104,6 @@ impl GameModeSettings {
         ui.spacing();
         ui.separator();
         new_settings
-    }
-}
-
-#[derive(Default, Reflect, PartialEq, Eq, Clone)]
-pub enum NewTabBehaviour {
-    Pop,
-    #[default]
-    SameNode,
-    SplitNode,
-}
-
-impl ToString for NewTabBehaviour {
-    fn to_string(&self) -> String {
-        match self {
-            Self::Pop => "New window",
-            Self::SameNode => "Same Node",
-            Self::SplitNode => "Splits Node",
-        }
-        .to_string()
-    }
-}
-
-#[derive(Default, Resource, Reflect)]
-#[reflect(Resource)]
-pub struct NewWindowSettings {
-    pub new_tab: NewTabBehaviour,
-}
-
-impl NewWindowSettings {
-    fn ui(&mut self, ui: &mut egui::Ui) {
-        egui::ComboBox::new("new_tab", "")
-            .selected_text(self.new_tab.to_string())
-            .show_ui(ui, |ui| {
-                for mode in TAB_MODES.into_iter() {
-                    if ui
-                        .selectable_label(self.new_tab == mode, mode.to_string())
-                        .clicked()
-                    {
-                        self.new_tab = mode;
-                    }
-                }
-            });
     }
 }
